requirements.txt
absl-py==0.4.0
appdirs==1.4.3
appnope==0.1.0
atomicwrites==1.2.0
attrs==18.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==1.5.0
cairocffi==0.9.0
CairoSVG==2.1.3
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
flake8==3.5.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
html5lib==0.9999999
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
line-profiler==2.1.2
lxml==4.2.4
Markdown==2.6.11
mccabe==0.6.1
more-itertools==4.3.0
numpy==1.15.1
opencv-python==3.4.2.17
parso==0.3.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pipdate==0.3.2
pluggy==0.7.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
py==1.5.4
pycodestyle==2.3.1
pycparser==2.18
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.7.3
python-chess==0.23.9
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
snakeviz==1.0.0
snowballstemmer==1.2.1
tensorflow==1.5.0
-e git+https://github.com/colwem/tensorflow_chessbot.git@6bf36072b770f4831cd0e7ad81ca800da0757ed8#egg=tensorflow_chessbot
tensorflow-tensorboard==1.5.1
tinycss2==0.6.1
tornado==5.1
traitlets==4.3.2
treelib==1.5.3
tuna==0.1.8
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
